Continuação melhorias de usabilidade

+++ app/src/main/java/br/edu/ifsp/scl/gerfinsdm2/adapter/ContaAdapter.kt
@@ -8,13 +8,9 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import br.edu.ifsp.scl.gerfinsdm2.R
 import br.edu.ifsp.scl.gerfinsdm2.activity.Conta.ContaDetalheActivity
-import br.edu.ifsp.scl.gerfinsdm2.data.ContaSQLite
-import br.edu.ifsp.scl.gerfinsdm2.data.TransacaoDAO
-import br.edu.ifsp.scl.gerfinsdm2.data.TransacaoSQLite
 import br.edu.ifsp.scl.gerfinsdm2.model.Conta
 import kotlinx.android.synthetic.main.conta_item.view.*
 import java.text.NumberFormat
-import java.text.SimpleDateFormat
 import java.util.*
 
 class ContaAdapter(private val context: Context, private var contaList: List<Conta>):
@@ -29,7 +25,7 @@ class ContaAdapter(private val context: Context, private var contaList: List<Con
         val textViewSaldo = itemView.tvSaldoConta
 
         fun bindView(conta: Conta) {
-            textViewSaldo.text = f.format(conta.saldoinicial)
+            textViewSaldo.text = f.format(conta.saldoinicial.toDouble())
             textViewNome.text = conta.nome
 
             // Tratamento do evento de click em um item da lista
